Gets Doneness History of User now
the database's "donenessHistory" is a map of String date keys to Integer arrays.
The first element of the array is the number of completed tasks, the second is the total number of tasks.

Also fixed a bug related to repetating tasks. (Had forgot to set it up to check if it repeated on that week day before. oops....)